test : `ApplicationTest` 코드 정리
- `IntegrationTestEnvironment` 로 환경설정 코드 정리
- `SimpleEntityManager` Test 코드 제거
